moved opengl, lang and config into api
