+++ amulet_map_editor/api/__init__.py
@@ -1,2 +1,3 @@
 from .version import version
 from .framework import AmuletUI
+from .logging import log
