+++ amulet_map_editor/api/opengl/mesh/__init__.py
@@ -1,0 +1,1 @@
+from amulet_map_editor.api.opengl.mesh.base.tri_mesh import TriMesh
