+++ amulet_map_editor/api/opengl/mesh/selection/render_selection.py
@@ -3,7 +3,7 @@ from OpenGL.GL import *
 import itertools
 from typing import Tuple, Dict, Any, Optional, Union
 
-from amulet_map_editor.opengl.mesh.base.tri_mesh import TriMesh
+from amulet_map_editor.api.opengl.mesh.base.tri_mesh import TriMesh
 from amulet.api.data_types import BlockCoordinatesAny, PointCoordinatesAny
