+++ amulet_map_editor/api/opengl/mesh/selection/render_selection_highlightable.py
