+++ amulet_map_editor/api/opengl/mesh/structure/structure.py
@@ -9,14 +9,14 @@ from amulet.api.structure import Structure
 from amulet.api.chunk import Chunk
 from amulet.api.block import BlockManager
 
-from amulet_map_editor.opengl.mesh.base.chunk_builder import RenderChunkBuilder
-from amulet_map_editor.opengl.resource_pack import ResourcePackManager
-from amulet_map_editor.opengl.mesh.selection import (
+from amulet_map_editor.api.opengl.mesh.base.chunk_builder import RenderChunkBuilder
+from amulet_map_editor.api.opengl.resource_pack import ResourcePackManager
+from amulet_map_editor.api.opengl.mesh.selection import (
     RenderSelectionGroup,
     RenderSelection,
 )
-from amulet_map_editor.opengl.mesh.base.tri_mesh import Drawable
-from amulet_map_editor.opengl.matrix import displacement_matrix
+from amulet_map_editor.api.opengl.mesh.base.tri_mesh import Drawable
+from amulet_map_editor.api.opengl.matrix import displacement_matrix
 
 
 class GreenRenderSelection(RenderSelection):
