+++ amulet_map_editor/programs/convert/convert.py
@@ -6,7 +6,8 @@ from typing import TYPE_CHECKING, Callable
 from amulet import world_interface
 from amulet.api.world import World
 
-from amulet_map_editor import lang, log
+from amulet_map_editor.api import lang
+from amulet_map_editor.api.logging import log
 from amulet_map_editor.amulet_wx.ui.simple import SimplePanel
 from amulet_map_editor.amulet_wx.ui.select_world import WorldSelectDialog, WorldUI
 from amulet_map_editor.api.datatypes import MenuData
